C# .Net Profesional 5.0 & 4.5.1
This book covers the .NET Framework 4.5.1.
If you code using any of the prior versions, there may be sections of the book that will not work for
you. This book notifies you of items that are new and specific to the .NET Framework 4.5 and 4.5.1.

This book can help you as reference in themes like NetWorking and WCF (Window Comunication Foundation)