MySql Development 5.1 by Sergei Golubchick & Andrew Hutchings
With this book you will learn how to develop MySql Plugins and UDF(User Defined Function) in C/C++
for Gnu/Linux and Mac and then call these function in your queries from MySql this will help you
create your avanced projects more easily.

The process is accompanied with a detailed step-by-step explanation of the code, at the same time providing you
with resources to create your own plugins.

If you are a MySQL application developer or a programmer who wants to interface with MySQL server,
this book is for you.

Basic knowledge of C/C++, SQL, and MySQL in particular is assumed in the book, but no deep familiarity with any of the
topics is required.