PHP & MySQL Novice To Ninja (Six Edition)
This book is aimed at intermediate and advanced web designers looking to
make leap into server-side programming.

By the end of this book, you can expect to have a grasp of what's involved
in building a modern PHP website, the basics of PHP, and tried and tested techniques
that are used by the developers.