Kali Linux 2 - Assuring Security by Pen Testting
Kali Linux is an operating system that has been designed to provide the penetration
tester a flexible platform to perform the panoply of penetration task such as enumerating a target,
identifying vulnerabilites, and exploting targets in a networked enviroment.

This volume presents a structured method for developing a skill set tailored to the unique
nature of penetration testing.