Android UI Design
Android is an OS based on the Linux Kernel and designed primarily for touchscreen
mobile devices such as smartphines and tablet computers.

Android Os has taken over the world of mobile devices and is growing rapidly the last few years.
Android interface is based on direct manipulation, using touch inputs that loosely correspond to
the real-worldactions, like swaping tapping, pinching and reverse piching to manipulate on-screen
objects.